templates/examples/custom_online_prediction/example_container/requirements.txt
joblib>=1.5.2
numpy>=2.3.3
scikit-learn>=1.7.2
google-cloud-storage>=2.19.0
tensorflow>=2.20.0
torch>=2.8.0
xgboost>=3.0.5
pandas>=2.3.2
